+++ src/repoproviders/resolvers/repos.py
@@ -4,6 +4,30 @@ from yarl import URL
 
 
 @dataclass(frozen=True)
+class Git:
+    repo: str
+    ref: str
+
+    immutable = False
+
+
+@dataclass(frozen=True)
+class ImmutableGit:
+    """
+    Same as Git, but marked to be fully resolved. This implies:
+
+    1. The repository exists, and can be contacted
+    2. If ref was a branch or tag, it has been resolved into an immutable commit sha
+    3. If ref *looks* like a sha, we assume it exists (without testing it)
+    """
+
+    repo: str
+    ref: str
+
+    immutable = True
+
+
+@dataclass(frozen=True)
 class GitHubURL:
     """
     A GitHub URL of any sort.
